fix: change addItem for veggies and co.

+++ src/main/java/com/scanly/scanlyBackend/dtos/UpdateItemQuantityRequest.java
@@ -1,5 +1,7 @@
 package com.scanly.scanlyBackend.dtos;
 
+import java.math.BigDecimal;
+
 public record UpdateItemQuantityRequest(
-        int delta
+        BigDecimal delta
 ){}
